refactor RoundTab from ProtocolDetails into molecules
