+++ app/src/organisms/ProtocolDetails/index.tsx
@@ -19,11 +19,9 @@ import {
   COLORS,
   DIRECTION_COLUMN,
   DIRECTION_ROW,
-  DISPLAY_BLOCK,
   DISPLAY_FLEX,
   JUSTIFY_CENTER,
   JUSTIFY_SPACE_BETWEEN,
-  POSITION_ABSOLUTE,
   POSITION_RELATIVE,
   SIZE_1,
   SIZE_5,
@@ -44,6 +42,7 @@ import { Divider } from '../../atoms/structure'
 import { StyledText } from '../../atoms/text'
 import { DeckThumbnail } from '../../molecules/DeckThumbnail'
 import { Modal } from '../../molecules/Modal'
+import { RoundTab } from '../../molecules/RoundTab'
 import { useTrackEvent } from '../../redux/analytics'
 import { getIsProtocolAnalysisInProgress } from '../../redux/protocol-storage'
 import { ChooseRobotToRunProtocolSlideout } from '../ChooseRobotToRunProtocolSlideout'
@@ -60,46 +59,6 @@ import { RobotConfigurationDetails } from './RobotConfigurationDetails'
 import type { JsonConfig, PythonConfig } from '@opentrons/shared-data'
 import type { StoredProtocolData } from '../../redux/protocol-storage'
 import type { State } from '../../redux/types'
-
-const defaultTabStyle = css`
-  ${TYPOGRAPHY.pSemiBold}
-  border-radius: ${BORDERS.radiusSoftCorners} ${BORDERS.radiusSoftCorners} 0 0;
-  border-top: ${BORDERS.transparentLineBorder};
-  border-left: ${BORDERS.transparentLineBorder};
-  border-right: ${BORDERS.transparentLineBorder};
-  padding: ${SPACING.spacing3} ${SPACING.spacing4};
-  position: ${POSITION_RELATIVE};
-`
-
-const inactiveTabStyle = css`
-  color: ${COLORS.darkGreyEnabled};
-
-  &:hover {
-    color: ${COLORS.darkGreyEnabled};
-    background-color: ${COLORS.fundamentalsBackgroundShade};
-  }
-`
-
-const currentTabStyle = css`
-  ${TYPOGRAPHY.pSemiBold}
-  background-color: ${COLORS.white};
-  border-top: ${BORDERS.lineBorder};
-  border-left: ${BORDERS.lineBorder};
-  border-right: ${BORDERS.lineBorder};
-  color: ${COLORS.blueEnabled};
-
-  /* extend below the tab when active to flow into the content */
-  &:after {
-    position: ${POSITION_ABSOLUTE};
-    display: ${DISPLAY_BLOCK};
-    content: '';
-    background-color: ${COLORS.white};
-    top: 100;
-    left: 0;
-    height: ${SIZE_1};
-    width: 100%;
-  }
-`
 
 const GRID_STYLE = css`
   display: grid;
@@ -122,34 +81,6 @@ const ZOOM_ICON_STYLE = css`
     box-shadow: 0 0 0 3px ${COLORS.fundamentalsFocus};
   }
 `
-
-interface RoundTabProps extends React.ComponentProps<typeof Btn> {
-  isCurrent: boolean
-}
-function RoundTab({
-  isCurrent,
-  children,
-  ...restProps
-}: RoundTabProps): JSX.Element {
-  return (
-    <Btn
-      {...restProps}
-      css={
-        isCurrent
-          ? css`
-              ${defaultTabStyle}
-              ${currentTabStyle}
-            `
-          : css`
-              ${defaultTabStyle}
-              ${inactiveTabStyle}
-            `
-      }
-    >
-      {children}
-    </Btn>
-  )
-}
 
 interface Metadata {
   [key: string]: any
